# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wabiz CRM API",
        "description": "Wabiz CRM - Multi-tenant WhatsApp CRM SaaS Platform API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/dashboard/overview": {
            "get": {
                "tags": [
                    "Dashboard Analytics"
                ],
                "summary": "Dashboard overview statistics",
                "description": "Get high-level KPIs: total leads, orders, revenue, messages, followups, conversion rate. All data scoped to the authenticated user.",
                "operationId": "d6549f214eb981b29935d8b2bd256aaa",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Preset date filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "today",
                                "yesterday",
                                "last_7_days",
                                "last_30_days",
                                "this_month",
                                "last_month",
                                "this_year"
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Custom start date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Custom end date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Overview statistics fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Overview statistics fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_leads": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "total_orders": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "total_revenue": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 12500.5
                                                },
                                                "total_messages": {
                                                    "type": "integer",
                                                    "example": 890
                                                },
                                                "total_followups": {
                                                    "type": "integer",
                                                    "example": 65
                                                },
                                                "pending_followups": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "delivered_orders": {
                                                    "type": "integer",
                                                    "example": 35
                                                },
                                                "conversion_rate": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 23.33
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}